requirements.txt
demucs==4.0.1
librosa==0.11.0
matplotlib==3.10.3
numpy==1.26.4
pytest==8.4.1
yt-dlp==2025.11.12
pydub==0.25.1
torchcodec==0.8.1